Output piano hint info for external use
I'll probably intend to use this information
with an Arduino and an LED strip to add "light"
to my keys in a couple of days. I know that the
author cautioned using that and that this
software is not supposed to teach this way.

Then again I love this software and it seems
Synthesia isn't opensource anymore anyway. So
I'll not ask the author to add this feature.
(I would've liked to add a hint midi output
anyway but not sure yet sure how to translate
good/bad or potential timings into useful
velocities and it's nothing big now anyway.)